Add the auto-disable feature to latency switching
To make the latency control safer, two new features have been
implemented:

1. To protect against possible misbehaving services, which crash while
holding the active latency request or just forget about releasing it -
the latency requests are now being disabled automatically after the
timeout. The timeout is fixed by 60 seconds, but it could be changed in
the future. Current and new latency control users should take it into
account and refresh their requests in time, according to the timeout
value.

2. To protect against collisions between identities used by different
callers, the client's `pid` value is added to the identity string. So if
two different services with pids 1111 and 2222 try to register similar
"latreq" requests, it will be transformed to the two different
identities:
- "1111:latreq"
- "2222:latreq"
To achieve this, the "clientPid" value is used from the payload, which
is added in ResSchedClient::ReportData.

Also organize the latency-related headers into the separate subdirectory.

Signed-off-by: Petrov Maxim <petrov.maxim@huawei.com>